html/semantics/embedded-content-0/media-elements/audio_defaultMuted_reflect.html

<!DOCTYPE html>
<html>
  <head>
    <title>Audio Test: audio_defaultMuted_check</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://dev.w3.org/html5/spec/single-page.html#dom-media-defaultmuted" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check that the defaultMuted reflects the muted content attribute in the audio element but has no dynamic effect" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <audio id="m" muted>The user agent doesn't support media element.</audio>
    <audio id="a">The user agent doesn't support media element.</audio>
    <script type="text/javascript">
        var media = document.getElementById("m");
        var audio = document.getElementById("a");
        
        test(function() {
            assert_true(media.defaultMuted, "media.defaultMuted of value");
        }, "Check if audio.defaultMuted returns true when the muted attribute is present in the audio element");
        
        test(function() {
            media.muted = false;
            assert_true(media.defaultMuted, "media.defaultMuted of value");
        }, "Check if audio.defaultMuted returns true after the muted attribute is set to false when the muted attribute is present in the audio element");
        
        test(function() {
            assert_false(audio.defaultMuted, "media.defaultMuted of value");
        }, "Check if audio.defaultMuted returns false when the muted attribute is absent in the audio element");
        
        test(function() {
            audio.muted = true;
            assert_false(audio.defaultMuted, "media.defaultMuted of value");
        }, "Check if audio.defaultMuted returns false after the muted attribute is set to true when the muted attribute is absent in the audio element");
    </script>
  </body>
</html>
